requirements.txt

flask==2.2.3
python-dotenv==1.0.0
docxtpl==0.16.7
pillow>=10.4.0
pywin32>=307; sys_platform == 'win32'
werkzeug==2.2.3
flask-wtf==1.1.1
python-docx
flask-login
flask-sqlalchemy
psycopg2-binary
beautifulsoup4
requests
itsdangerous
